Update pull request workflow and improve form labels
Refactor the GitHub Actions workflow to ensure proper setup of Node.js and installation of dependencies. Additionally, update the labels in the login and register forms to use lowercase for consistency. New entries for email and password have also been added to the localization file for internationalization support.

Issue: MW-0000
Type: chore