Rename workflow from Code quality to Lint
This change updates the workflow name to better reflect its purpose. The previous name was misleading, and renaming it to 'Lint' clarifies that this workflow is focused on linting tasks. Additionally, the branch specification for the push event has been removed to allow for broader applicability.

Issue: MW-0000
Type: chore